add TYPE_DECLARATOIN_STRICT set to separate strict and normal types
